Update 01-introduction.md
Cut out lots of content to simplify this introduction. Working towards an introduction session of teaching time about 15 minutes long, not including the exercise at the end.  Motivation behind the structure: Start with the context, outline the key principles and provide a couple of resources to bookmark for later.  Still want to revise this content again - some claims need verifying.