Update 08-software.md
This first draft of the lesson provides a summary of how FAIR can be applied to software. For structure and readability, the lesson is organized into sub-sections for each of the principles. 

The current draft is based on the Top 10 FAIR Data & Software Things (https://librarycarpentry.org/Top-10-FAIR//2018/12/01/research-software/) & FAIR Software guide (https://fair-software.eu/). It should provide a structure and foundation on which to expand.

At the moment, links and examples to support the content (as well as additional resources) are missing and will be updated soon!